task4/project-root/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
